+++ cypress/integration/SmokeTests/LoginWithUserEnglish/index.js
@@ -1,13 +1,14 @@
-beforeEach(() => {
-    cy.loginThroughForm('dhis2UsernameEnglish', 'dhis2PasswordEnglish');
-});
+// Blocked by DHIS2-13287
+// beforeEach(() => {
+//     cy.loginThroughForm('dhis2UsernameEnglish', 'dhis2PasswordEnglish');
+// });
 
-Given('you open the capture app', () => {
-    cy.visit('/');
-});
+// Given('you open the capture app', () => {
+//     cy.visit('/');
+// });
 
-Then('you should see the app main selections', () => {
-    cy.get('[data-test="org-unit-selector-container"]');
+// Then('you should see the app main selections', () => {
+//     cy.get('[data-test="org-unit-selector-container"]');
 
-    cy.get('[data-test="program-selector-container"]');
-});
+//     cy.get('[data-test="program-selector-container"]');
+// });