+++ src/core_modules/capture-core-utils/rulesEngine/constants/environmentTypes.const.js
@@ -1,0 +1,6 @@
+// @flow
+export const environmentTypes = Object.freeze({
+    WebClient: 'WebClient',
+    AndroidClient: 'AndroidClient',
+    Server: 'Server',
+});