+++ src/core_modules/capture-core-utils/rulesEngine/constants/index.js
@@ -4,3 +4,4 @@ export { rulesEngineEffectTargetDataTypes } from './targetDataTypes.const';
 export { idNames } from './idNames.const';
 export { typeKeys } from './typeKeys.const';
 export { mapTypeToInterfaceFnName } from './typeToInterfaceFnName.const';
+export { environmentTypes } from './environmentTypes.const';