+++ src/core_modules/capture-ui/internal/SelectionBoxes/SingleSelectionBoxes/SingleSelectBox/SingleSelectBox.component.js
@@ -81,7 +81,6 @@ export class SingleSelectBox extends React.Component<Props> {
                         id={id}
                         checked={isSelected}
                         value={optionData.value}
-                        onChange={this.handleSelect}
                         onClick={this.handleSelect}
                         onKeyPress={this.handleKeyPress}
                         disabled={disabled}