+++ CHANGELOG.md
@@ -1,3 +1,65 @@
+## [100.3.2](https://github.com/dhis2/capture-app/compare/v100.3.1...v100.3.2) (2022-05-10)
+
+
+### Bug Fixes
+
+* [DHIS2-13241] URL out of sync after adding event ([66317d4](https://github.com/dhis2/capture-app/commit/66317d4e5615d516bc25c32f4c14e29f1ad316a1))
+
+## [100.3.1](https://github.com/dhis2/capture-app/compare/v100.3.0...v100.3.1) (2022-05-05)
+
+
+### Bug Fixes
+
+* initial CD test ([cea359b](https://github.com/dhis2/capture-app/commit/cea359b64dba0388bfa46872c7343479e9ee76a6))
+
+# [100.3.0](https://github.com/dhis2/capture-app/compare/v100.2.4...v100.3.0) (2022-05-05)
+
+
+### Features
+
+* [DHIS2-12246] add missing built-in variables in the rules engine ([#2663](https://github.com/dhis2/capture-app/issues/2663)) ([4cb18bc](https://github.com/dhis2/capture-app/commit/4cb18bc347615e8b305aa63c282282c1fa09efba))
+* ignore diacritics in option group search ([#2678](https://github.com/dhis2/capture-app/issues/2678)) ([c33a660](https://github.com/dhis2/capture-app/commit/c33a66001742d91e0d8391ecfe7581294a934a72))
+
+## [100.2.4](https://github.com/dhis2/capture-app/compare/v100.2.3...v100.2.4) (2022-05-02)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([42350cd](https://github.com/dhis2/capture-app/commit/42350cdd56cfdbee6e90a07ca85507e2fa67df28))
+
+## [100.2.3](https://github.com/dhis2/capture-app/compare/v100.2.2...v100.2.3) (2022-04-25)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([2cf12aa](https://github.com/dhis2/capture-app/commit/2cf12aa0d9760158a2fb6587c5797bc8c17ac201))
+
+## [100.2.2](https://github.com/dhis2/capture-app/compare/v100.2.1...v100.2.2) (2022-04-22)
+
+
+### Bug Fixes
+
+* [TECH-1077] remove unnecessary onChange handler in SingleSelectBox ([6a5eab0](https://github.com/dhis2/capture-app/commit/6a5eab0b2cb04e067d49ba9294899d177b92f50c))
+
+## [100.2.1](https://github.com/dhis2/capture-app/compare/v100.2.0...v100.2.1) (2022-04-22)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([ba0564a](https://github.com/dhis2/capture-app/commit/ba0564acc7ec2f76cf9f79a3b8dff0bef6f34860))
+
+# [100.2.0](https://github.com/dhis2/capture-app/compare/v100.1.15...v100.2.0) (2022-04-21)
+
+
+### Bug Fixes
+
+* [DHIS2-12675] custom relative range for date ([#2646](https://github.com/dhis2/capture-app/issues/2646)) ([68e01a4](https://github.com/dhis2/capture-app/commit/68e01a403cd0ce1d12ca30644f01c9089724969f))
+
+
+### Features
+
+* [DHIS2-12885] Opt in and out of using the Enrollment dashboard ([#2703](https://github.com/dhis2/capture-app/issues/2703)) ([ccdffdf](https://github.com/dhis2/capture-app/commit/ccdffdf0c03d0dff17df8dc99239e1a7614a2970))
+
 ## [100.1.15](https://github.com/dhis2/capture-app/compare/v100.1.14...v100.1.15) (2022-04-20)